refactor(ticker): move result to an undefined interface
